main/server/requirements.txt
# Core packages
crewai==0.1.25
langchain>=0.1.0
langchain-community>=0.0.10
pydantic==2.5.3

# LLM integration
openai>=1.3.0
anthropic>=0.5.0
python-dotenv==1.0.0

# Template and document processing
jinja2>=3.1.2
python-docx>=0.8.11
markdown>=3.4.3
PyPDF2==3.0.1  # Updated to specific version
fpdf==1.7.2

# NLP and text analysis
spacy>=3.6.0
nltk>=3.8.1

# Data handling and storage
neo4j>=5.9.0
pymongo>=4.5.0

# Utilities
PyYAML>=6.0
python-dateutil>=2.8.2
tqdm>=4.65.0

# Logging and monitoring
loguru>=0.7.0

# Development and testing
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0

# Web framework
fastapi==0.109.1
uvicorn==0.27.0